// common/types.ts
export const PlankaTaskListSchema = z.object({
  id: z.string(),
  cardId: z.string(),
  name: z.string(),
  position: z.number(),
  showOnFrontOfCard: z.boolean().optional(),
  hideCompletedTasks: z.boolean().optional(),
  createdAt: z.string().optional(),
  updatedAt: z.string().nullable().optional(),
});

export const PlankaTaskSchema = z.object({
  id: z.string(),
  taskListId: z.string().optional(),
  cardId: z.string().optional(),
  name: z.string(),
  isCompleted: z.boolean().optional(),
  position: z.number(),
  assigneeUserId: z.string().nullable().optional(),
  createdAt: z.string().optional(),
  updatedAt: z.string().nullable().optional(),
});